+++ forge/__init__.py
@@ -17,7 +17,7 @@ class Application(object):
         self.registry = {}
 
         plugins = self.parse_conf('/usr/local/etc/forge/', 'conf.ini')
-        plugins.append('/usr/local/etc/forge/plugins/add_plugin')
+        plugins.append('/usr/local/etc/forge/plugins/manage_plugins')
         print(plugins)
         self.plugin_source = plugin_base.make_plugin_source(
             searchpath=plugins,
